// builder/builder_types/enums.ts
export enum ComponentLocations {
  DEADLINE__COMPONENT = "DEADLINE__COMPONENT",
  PURCHASE__COMPONENT = "PURCHASE__COMPONENT",
  SIGNATURE__COMPONENT = "SIGNATURE__COMPONENT",
  BROKERS__ACKNOWLEDGMENTS = "BROKERS__ACKNOWLEDGMENTS",
}

export enum SpacePropertyKeys {
  BUYER_KEY = "Buyer",
  SELLER_KEY = "Seller",
  BUYERS_AGENT_KEY = "BUYERS_AGENT_KEY",
  SELLERS_AGENT_KEY = "SELLERS_AGENT_KEY",
  PROPERTY_KEY = "Property",
}

export enum Inputs {
  TEXT__BOX__SMALL = "TEXT__BOX__SMALL",
  CHECKBOX = "CHECKBOX",
  FLOATING_CHECKBOX = "FLOATING_CHECKBOX",
  RADIO = "RADIO",
  ADDRESS_BLOCK = "ADDRESS_BLOCK",
  TEXT__BOX__LARGE = "TEXT__BOX__LARGE",
  TEXT__BOX__MEDIUM = "TEXT__BOX__MEDIUM",
  DATE__BOX = "DATE__BOX",
  NUMBER_INPUT = "NUMBER_INPUT",

  FULL__LENGTH__PARAGRAPH__2 = "FULL__LENGTH__PARAGRAPH__2",
  FULL__LENGTH__PARAGRAPH__3 = "FULL__LENGTH__PARAGRAPH__3",
  FULL__LENGTH__PARAGRAPH__5 = "FULL__LENGTH__PARAGRAPH__5",
  FULL__LENGTH__PARAGRAPH__4 = "FULL__LENGTH__PARAGRAPH__4",
  FULL__LENGTH__PARAGRAPH__7 = "FULL__LENGTH__PARAGRAPH__7",
  FULL__LENGTH__PARAGRAPH__10 = "FULL__LENGTH__PARAGRAPH__10",
  SKIP__ROW = "SKIP__ROW",
  TEXT_REPLACE = "TEXT_REPLACE",
}

export enum RowStyles {
  CENTERED = "CENTERED",
  ALIGNED__TOP = "ALIGNED__TOP",
  ALIGNED__RIGHT = "ALIGNED__RIGHT",
  ALIGNED__LEFT = "ALIGNED__LEFT",
  FONT__SIZE__LARGE = "FONT__SIZE__LARGE",
  FONT__SIZE__LARGE__XL = "FONT__SIZE__LARGE__XL",
  ZERO__PADDING__LEFT = "ZERO__PADDING__LEFT",
  ZERO__PADDING__RIGHT = "ZERO__PADDING__RIGHT",
  BORDER__TOP = "BORDER__TOP",
  BORDER__BOTTOM = "BORDER__BOTTOM",
  BORDER__LEFT = "BORDER__LEFT",
  BORDER__RIGHT = "BORDER__RIGHT",
  FORCE_FULL_WIDTH = "FORCE_FULL_WIDTH",
  FIT_WIDTH = "FIT_WIDTH",
  INVISIBLE = "INVISIBLE",
  LINE__HEIGHT__LG = "LINE__HEIGHT__LG",
  INNER_BORDER = "INNER_BORDER",
}

export enum SegmentStyles {
  BOLD__SECTION__TITLES__TO__FIRST__PERIOD = "BOLD__SECTION__TITLES__TO__FIRST__PERIOD",
  BOLD__ALL = "BOLD__ALL",
  BOLD__NOTE_AND_NOT = "BOLD__NOTE_AND_NOT",
  BOLD__UP_TO_BRACKET_AND_NOTE = "BOLD__UP_TO_BRACKET_AND_NOTE",

  PATTERN__REPLACE = "PATTERN__REPLACE",
}

export enum StaticStyles {
  IGNORE_HEADER = "IGNORE_HEADER",
  IGNORE_TITLE = "IGNORE_TITLE",
  KEEP__SINGLE = "KEEP__SINGLE",
  DEADLINE__COMPONENT = "DEADLINE__COMPONENT",
  LINE_NUMBER_UNSET = "LINE_NUMBER_UNSET",
}